backend-structure - Moved services and it's exception into a 'Services' folder
