+++ app/Http/Controllers/API/FolderController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\API;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-use App\QuestionService\Contracts\QuestionServiceInterface;
+use App\Services\Questions\Contracts\QuestionServiceInterface;
 use Illuminate\Support\Facades\Response;
 
 class FolderController extends Controller
