+++ app/Providers/AuthServiceProvider.php
@@ -24,8 +24,8 @@ class AuthServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->singleton(
-            'App\AuthService\Contracts\AuthServiceInterface',
-            'App\AuthService\AuthService'
+            'App\Services\Auth\Contracts\AuthServiceInterface',
+            'App\Services\Auth\AuthService'
         );
     }
 }
