+++ app/Repositories/Exceptions/RepositoryException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Exceptions;
+namespace App\Repositories\Exceptions;
 
 class RepositoryException extends \RuntimeException
 {}
